working MemSink
