+++ cvmfs/network/download.h
@@ -22,6 +22,8 @@
 #include "crypto/hash.h"
 #include "dns.h"
 #include "duplex_curl.h"
+#include "jobinfo.h"
+#include "network_errors.h"
 #include "sink.h"
 #include "ssl.h"
 #include "statistics.h"
@@ -33,98 +35,6 @@
 class InterruptCue;
 
 namespace download {
-
-/**
- * Possible return values.  Adjust ObjectFetcher error handling if new network
- * error conditions are added.
- */
-enum Failures {
-  kFailOk = 0,
-  kFailLocalIO,
-  kFailBadUrl,
-  kFailProxyResolve,
-  kFailHostResolve,
-  // artificial failure code.  Try other host even though
-  // failure seems to be at the proxy
-  kFailHostAfterProxy,
-  kFailProxyConnection,
-  kFailHostConnection,
-  kFailProxyHttp,
-  kFailHostHttp,
-  kFailBadData,
-  kFailTooBig,
-  kFailOther,
-  kFailUnsupportedProtocol,
-  kFailProxyTooSlow,
-  kFailHostTooSlow,
-  kFailProxyShortTransfer,
-  kFailHostShortTransfer,
-  kFailCanceled,
-
-  kFailNumEntries
-};  // Failures
-
-
-inline bool IsHostTransferError(const Failures error) {
-  switch (error) {
-    case kFailHostConnection:
-    case kFailHostTooSlow:
-    case kFailHostShortTransfer:
-      return true;
-    default:
-      break;
-  }
-  return false;
-}
-
-inline bool IsProxyTransferError(const Failures error) {
-  switch (error) {
-    case kFailProxyConnection:
-    case kFailProxyTooSlow:
-    case kFailProxyShortTransfer:
-      return true;
-    default:
-      break;
-  }
-  return false;
-}
-
-inline const char *Code2Ascii(const Failures error) {
-  const char *texts[kFailNumEntries + 1];
-  texts[0] = "OK";
-  texts[1] = "local I/O failure";
-  texts[2] = "malformed URL";
-  texts[3] = "failed to resolve proxy address";
-  texts[4] = "failed to resolve host address";
-  texts[5] = "all proxies failed, trying host fail-over";
-  texts[6] = "proxy connection problem";
-  texts[7] = "host connection problem";
-  texts[8] = "proxy returned HTTP error";
-  texts[9] = "host returned HTTP error";
-  texts[10] = "corrupted data received";
-  texts[11] = "resource too big to download";
-  texts[12] = "unknown network error";
-  texts[13] = "Unsupported URL in protocol";
-  texts[14] = "proxy serving data too slowly";
-  texts[15] = "host serving data too slowly";
-  texts[16] = "proxy data transfer cut short";
-  texts[17] = "host data transfer cut short";
-  texts[18] = "request canceled";
-  texts[19] = "no text";
-  return texts[error];
-}
-
-/**
- * Where to store downloaded data.
- */
-enum Destination {
-  kDestinationMem = 1,
-  kDestinationFile,
-  kDestinationPath,
-  kDestinationSink,
-  kDestinationNone
-};  // Destination
-
 
 struct Counters {
   perf::Counter *sz_transferred_bytes;
@@ -148,162 +58,6 @@ struct Counters {
         "Number of host failovers");
   }
 };  // Counters
-
-
-/**
- * Contains all the information to specify a download job.
- */
-struct JobInfo {
-  const std::string *url;
-  bool compressed;
-  bool probe_hosts;
-  bool head_request;
-  bool follow_redirects;
-  bool force_nocache;
-  pid_t pid;
-  uid_t uid;
-  gid_t gid;
-  void *cred_data;  // Per-transfer credential data
-  InterruptCue *interrupt_cue;
-  Destination destination;
-  struct {
-    size_t size;
-    size_t pos;
-    char *data;
-  } destination_mem;
-  FILE *destination_file;
-  const std::string *destination_path;
-  cvmfs::Sink *destination_sink;
-  const shash::Any *expected_hash;
-  const std::string *extra_info;
-
-  // Allow byte ranges to be specified.
-  off_t range_offset;
-  off_t range_size;
-
-  // Default initialization of fields
-  void Init() {
-    url = NULL;
-    compressed = false;
-    probe_hosts = false;
-    head_request = false;
-    follow_redirects = false;
-    force_nocache = false;
-    pid = -1;
-    uid = -1;
-    gid = -1;
-    cred_data = NULL;
-    interrupt_cue = NULL;
-    destination = kDestinationNone;
-    destination_mem.size = destination_mem.pos = 0;
-    destination_mem.data = NULL;
-    destination_file = NULL;
-    destination_path = NULL;
-    destination_sink = NULL;
-    expected_hash = NULL;
-    extra_info = NULL;
-
-    curl_handle = NULL;
-    headers = NULL;
-    memset(&zstream, 0, sizeof(zstream));
-    info_header = NULL;
-    pipe_job_results = NULL;
-    nocache = false;
-    error_code = kFailOther;
-    num_used_proxies = num_used_hosts = num_retries = 0;
-    backoff_ms = 0;
-    current_host_chain_index = 0;
-
-    range_offset = -1;
-    range_size = -1;
-    http_code = -1;
-  }
-
-  // One constructor per destination + head request
-  JobInfo() { Init(); }
-  JobInfo(const std::string *u, const bool c, const bool ph,
-          const std::string *p, const shash::Any *h)
-  {
-    Init();
-    url = u;
-    compressed = c;
-    probe_hosts = ph;
-    destination = kDestinationPath;
-    destination_path = p;
-    expected_hash = h;
-  }
-  JobInfo(const std::string *u, const bool c, const bool ph, FILE *f,
-          const shash::Any *h)
-  {
-    Init();
-    url = u;
-    compressed = c;
-    probe_hosts = ph;
-    destination = kDestinationFile;
-    destination_file = f;
-    expected_hash = h;
-  }
-  JobInfo(const std::string *u, const bool c, const bool ph,
-          const shash::Any *h)
-  {
-    Init();
-    url = u;
-    compressed = c;
-    probe_hosts = ph;
-    destination = kDestinationMem;
-    expected_hash = h;
-  }
-  JobInfo(const std::string *u, const bool c, const bool ph,
-          cvmfs::Sink *s, const shash::Any *h)
-  {
-    Init();
-    url = u;
-    compressed = c;
-    probe_hosts = ph;
-    destination = kDestinationSink;
-    destination_sink = s;
-    expected_hash = h;
-  }
-  JobInfo(const std::string *u, const bool ph) {
-    Init();
-    url = u;
-    probe_hosts = ph;
-    head_request = true;
-  }
-
-  ~JobInfo() {
-    if (pipe_job_results.IsValid()) {
-      pipe_job_results.Destroy();
-    }
-  }
-
-  /**
-   * Tells whether the error is because of a non-existing file. Should only
-   * be called if error_code is not kFailOk
-   */
-  bool IsFileNotFound();
-
-  // Internal state, don't touch
-  CURL *curl_handle;
-  curl_slist *headers;
-  char *info_header;
-  z_stream zstream;
-  shash::ContextPtr hash_context;
-
-  /// Pipe used for the return value
-  UniquePtr<Pipe<kPipeDownloadJobsResults> > pipe_job_results;
-
-  std::string proxy;
-  bool nocache;
-  Failures error_code;
-  int http_code;
-  unsigned char num_used_proxies;
-  unsigned char num_used_hosts;
-  unsigned char num_retries;
-  unsigned backoff_ms;
-  unsigned int current_host_chain_index;
-};  // JobInfo
-
 
 /**
  * Manages blocks of arrays of curl_slist storing header strings.  In contrast
