+++ cvmfs/swissknife_info.cc
@@ -127,18 +127,19 @@ int swissknife::CommandInfo::Main(const swissknife::ArgumentList &args) {
   UniquePtr<manifest::Manifest> manifest;
   if (IsRemote(repository)) {
     const string url = repository + "/.cvmfspublished";
-    download::JobInfo download_manifest(&url, false, false, NULL);
+    cvmfs::MemSink memsink;
+    download::JobInfo download_manifest(&url, false, false, NULL, &memsink);
     download::Failures retval = download_manager()->Fetch(&download_manifest);
     if (retval != download::kFailOk) {
       LogCvmfs(kLogCvmfs, kLogStderr, "failed to download manifest (%d - %s)",
                retval, download::Code2Ascii(retval));
       return 1;
     }
-    char *buffer = download_manifest.destination_mem.data;
-    const unsigned length = download_manifest.destination_mem.pos;
+    char *buffer = download_manifest.destination_memsink->data_;
+    const unsigned length = download_manifest.destination_memsink->pos_;
     manifest = manifest::Manifest::LoadMem(
         reinterpret_cast<const unsigned char *>(buffer), length);
-    free(download_manifest.destination_mem.data);
+    download_manifest.destination_memsink->Reset();
   } else {
     if (chdir(repository.c_str()) != 0) {
       LogCvmfs(kLogCvmfs, kLogStderr, "failed to switch to directory %s",
@@ -250,7 +251,8 @@ int swissknife::CommandInfo::Main(const swissknife::ArgumentList &args) {
       return 0;
     }
     const string url = repository + "/data/" + meta_info.MakePath();
-    download::JobInfo download_metainfo(&url, true, false, &meta_info);
+    cvmfs::MemSink memsink;
+    download::JobInfo download_metainfo(&url, true, false, &meta_info, &memsink);
     download::Failures retval = download_manager()->Fetch(&download_metainfo);
     if (retval != download::kFailOk) {
       if (human_readable)
@@ -259,8 +261,8 @@ int swissknife::CommandInfo::Main(const swissknife::ArgumentList &args) {
                  download::Code2Ascii(retval));
       return 1;
     }
-    string info(download_metainfo.destination_mem.data,
-                download_metainfo.destination_mem.pos);
+    string info(download_metainfo.destination_memsink->data_,
+                download_metainfo.destination_memsink->pos_);
     LogCvmfs(kLogCvmfs, kLogStdout | kLogNoLinebreak, "%s", info.c_str());
   }
