+++ test/unittests/CMakeLists.txt
@@ -179,6 +179,7 @@ if (BUILD_SERVER)
                   ${CVMFS_SOURCE_DIR}/manifest_fetch.cc
                   ${CVMFS_SOURCE_DIR}/network/dns.cc
                   ${CVMFS_SOURCE_DIR}/network/download.cc
+                  ${CVMFS_SOURCE_DIR}/network/jobinfo.cc
                   ${CVMFS_SOURCE_DIR}/network/s3fanout.cc
                   ${CVMFS_SOURCE_DIR}/options.cc
                   ${CVMFS_SOURCE_DIR}/pack.cc
@@ -424,6 +425,7 @@ set(CVMFS_UNITTEST_SOURCES
   ${CVMFS_SOURCE_DIR}/mountpoint.cc
   ${CVMFS_SOURCE_DIR}/network/dns.cc
   ${CVMFS_SOURCE_DIR}/network/download.cc
+  ${CVMFS_SOURCE_DIR}/network/jobinfo.cc
   ${CVMFS_SOURCE_DIR}/network/s3fanout.cc
   ${CVMFS_SOURCE_DIR}/notify/messages.cc
   ${CVMFS_SOURCE_DIR}/options.cc
